@@ -1,10 +1,10 @@
 [workspace.package]
-version = "0.5.1"
+version = "0.5.2"
 authors = [
     "William Brown <william@blackhats.net.au>",
     "Michael Farrell <micolous+git@gmail.com>",
 ]
-rust-version = "1.81"
+rust-version = "1.85"
 edition = "2021"
 repository = "https://github.com/kanidm/webauthn-rs"
 homepage = "https://github.com/kanidm/webauthn-rs"
@@ -56,14 +56,14 @@ exclude = ["compat_tester/webauthn-rs-demo-wasm", "tutorial/wasm"]
 
 [workspace.dependencies]
 # These are in release/dependency order.
-base64urlsafedata = { path = "./base64urlsafedata", version = "=0.5.1" }
-fido-hid-rs = { path = "./fido-hid-rs", version = "=0.5.1" }
-webauthn-attestation-ca = { path = "./attestation-ca", version = "=0.5.1" }
-webauthn-rs-proto = { path = "./webauthn-rs-proto", version = "=0.5.1" }
-fido-mds = { path = "./fido-mds", version = "=0.5.1" }
-webauthn-rs-core = { path = "./webauthn-rs-core", version = "=0.5.1" }
-webauthn-rs = { path = "./webauthn-rs", version = "=0.5.1" }
-webauthn-authenticator-rs = { path = "./webauthn-authenticator-rs", version = "=0.5.1" }
+base64urlsafedata = { path = "./base64urlsafedata", version = "=0.5.2" }
+fido-hid-rs = { path = "./fido-hid-rs", version = "=0.5.2" }
+webauthn-attestation-ca = { path = "./attestation-ca", version = "=0.5.2" }
+webauthn-rs-proto = { path = "./webauthn-rs-proto", version = "=0.5.2" }
+fido-mds = { path = "./fido-mds", version = "=0.5.2" }
+webauthn-rs-core = { path = "./webauthn-rs-core", version = "=0.5.2" }
+webauthn-rs = { path = "./webauthn-rs", version = "=0.5.2" }
+webauthn-authenticator-rs = { path = "./webauthn-authenticator-rs", version = "=0.5.2" }
 
 # Currently un-released
 cable-tunnel-server-common = { path = "./cable-tunnel-server/common", version = "0.1.0" }
@@ -85,6 +85,7 @@ nom = "7.1"
 num-derive = { version = "0.4.2" }
 peg = "0.8.1"
 openssl = "^0.10.56"
+openssl-sys = "^0.9.109"
 
 rand = "0.8"
 rand_chacha = "0.3.1"
