@@ -8,10 +8,7 @@ use crate::crypto::{
 use crate::error::WebauthnError;
 use crate::internals::*;
 use crate::proto::*;
-use base64urlsafedata::HumanBinaryData;
-use compact_jwt::{crypto::JwsX509Verifier, JwsCompact, JwsVerifier};
 use openssl::hash::MessageDigest;
-use openssl::sha::sha256;
 use openssl::stack;
 use openssl::x509;
 use openssl::x509::store;
@@ -31,7 +28,7 @@ pub trait AttestationX509Extension {
     const OID: Oid<'static>;
 
     /// how to parse the value out of the certificate extension
-    fn parse(i: &[u8]) -> der_parser::error::BerResult<(Self::Output, AttestationMetadata)>;
+    fn parse(i: &[u8]) -> der_parser::error::BerResult<'_, (Self::Output, AttestationMetadata)>;
 
     /// if `true`, then validating this certificate fails if this extension is
     /// missing
@@ -55,7 +52,7 @@ impl AttestationX509Extension for FidoGenCeAaguid {
     // verify that the value of this extension matches the aaguid in authenticatorData.
     type Output = Aaguid;
 
-    fn parse(i: &[u8]) -> der_parser::error::BerResult<(Self::Output, AttestationMetadata)> {
+    fn parse(i: &[u8]) -> der_parser::error::BerResult<'_, (Self::Output, AttestationMetadata)> {
         let (rem, aaguid) = der_parser::der::parse_der_octetstring(i)?;
         let aaguid: Aaguid = aaguid
             .as_slice()
@@ -116,7 +113,7 @@ pub(crate) mod android_key_attestation {
     }
 
     impl AuthorizationList {
-        pub fn parse(i: &[u8]) -> der_parser::error::BerResult<Self> {
+        pub fn parse(i: &[u8]) -> der_parser::error::BerResult<'_, Self> {
             use der_parser::{der::*, error::BerError};
             parse_der_container(|i: &[u8], hdr: Header| {
                 if hdr.tag() != Tag::Sequence {
@@ -168,7 +165,7 @@ pub(crate) mod android_key_attestation {
     }
 
     impl Data {
-        pub fn parse(i: &[u8]) -> der_parser::error::BerResult<(Vec<u8>, AttestationMetadata)> {
+        pub fn parse(i: &[u8]) -> der_parser::error::BerResult<'_, (Vec<u8>, AttestationMetadata)> {
             use der_parser::{der::*, error::BerError};
             parse_der_container(|i: &[u8], hdr: Header| {
                 if hdr.tag() != Tag::Sequence {
@@ -267,7 +264,7 @@ impl AttestationX509Extension for AndroidKeyAttestationExtensionData {
     // verify that the value of this extension matches the aaguid in authenticatorData.
     type Output = Vec<u8>;
 
-    fn parse(i: &[u8]) -> der_parser::error::BerResult<(Self::Output, AttestationMetadata)> {
+    fn parse(i: &[u8]) -> der_parser::error::BerResult<'_, (Self::Output, AttestationMetadata)> {
         android_key_attestation::Data::parse(i)
     }
 
@@ -282,7 +279,7 @@ impl AttestationX509Extension for AppleAnonymousNonce {
     // 4. Verify that nonce equals the value of the extension with OID ( 1.2.840.113635.100.8.2 ) in credCert. The nonce here is used to prove that the attestation is live and to protect the integrity of the authenticatorData and the client data.
     const OID: Oid<'static> = der_parser::oid!(1.2.840 .113635 .100 .8 .2);
 
-    fn parse(i: &[u8]) -> der_parser::error::BerResult<(Self::Output, AttestationMetadata)> {
+    fn parse(i: &[u8]) -> der_parser::error::BerResult<'_, (Self::Output, AttestationMetadata)> {
         use der_parser::{der::*, error::BerError};
         parse_der_container(|i: &[u8], hdr: Header| {
             if hdr.tag() != Tag::Sequence {
@@ -1195,181 +1192,6 @@ pub(crate) fn verify_android_key_attestation(
 
     // 5. If successful, return implementation-specific values representing attestation type Anonymous CA and attestation trust path x5c.
     Ok((ParsedAttestationData::Basic(arr_x509), meta))
-}
-
-/// <https://www.w3.org/TR/webauthn/#sctn-android-safetynet-attestation>
-pub(crate) fn verify_android_safetynet_attestation(
-    _acd: &AttestedCredentialData,
-    att_obj: &AttestationObject<Registration>,
-    client_data_hash: &[u8],
-    danger_ignore_timestamp: bool,
-) -> Result<(ParsedAttestationData, AttestationMetadata), WebauthnError> {
-    let att_stmt = &att_obj.att_stmt;
-    let auth_data_bytes = &att_obj.auth_data_bytes;
-
-    // 1. Verify that attStmt is valid CBOR conforming to the syntax defined above and perform CBOR decoding on it to extract the contained fields.
-    let att_stmt_map =
-        cbor_try_map!(att_stmt).map_err(|_| WebauthnError::AttestationStatementMapInvalid)?;
-
-    // there's only 1 version now
-    let _ver = {
-        let ver = att_stmt_map
-            .get(&serde_cbor_2::Value::Text("ver".to_string()))
-            .ok_or(WebauthnError::AttestationStatementVerMissing)?;
-
-        cbor_try_string!(ver).map_err(|_| WebauthnError::AttestationStatementVerInvalid)?
-    };
-
-    let response = {
-        let response = att_stmt_map
-            .get(&serde_cbor_2::Value::Text("response".to_string()))
-            .ok_or(WebauthnError::AttestationStatementResponseMissing)?;
-
-        cbor_try_bytes!(response).map_err(|_| WebauthnError::AttestationStatementResponseMissing)?
-    };
-
-    // Concatenate authenticatorData and clientDataHash to form the data to verify.
-    let data_to_verify: Vec<u8> = auth_data_bytes
-        .iter()
-        .chain(client_data_hash.iter())
-        .copied()
-        .collect();
-    let data_to_verify = sha256(&data_to_verify);
-
-    // 2. Verify that response is a valid SafetyNet response of version ver by following the steps
-    // indicated by the SafetyNet online documentation. As of this writing, there is only one format
-    // of the SafetyNet response and ver is reserved for future use.
-    #[derive(Clone, Debug, serde::Deserialize, serde::Serialize)]
-    #[serde(rename_all = "camelCase")]
-    struct SafteyNetAttestResponse {
-        timestamp_ms: u64,
-        nonce: Base64UrlSafeData,
-        apk_package_name: String,
-        apk_certificate_digest_sha256: Vec<HumanBinaryData>,
-        cts_profile_match: bool,
-        basic_integrity: bool,
-        evaluation_type: Option<String>,
-    }
-
-    let response_str = std::str::from_utf8(response.as_slice())
-        .map_err(|_| WebauthnError::AttestationStatementResponseInvalid)?;
-
-    #[derive(Debug, thiserror::Error)]
-    #[allow(missing_docs)]
-    enum SafetyNetError {
-        #[error("JWT error")]
-        Jwt(#[from] compact_jwt::JwtError),
-
-        #[error("No cert in chain")]
-        MissingCertChain,
-
-        #[error("Invalid Cert")]
-        BadCert,
-
-        #[error("Base64 error: {0}")]
-        Base64(#[from] base64::DecodeError),
-
-        #[error("openssl")]
-        OpenSSL(#[from] openssl::error::ErrorStack),
-
-        #[error("nonce mismatch")]
-        NonceMismatch,
-
-        #[error("hostname invalid")]
-        InvalidHostname,
-
-        #[error("False CTS Profile Match")]
-        CtsProfileMatchFailed,
-
-        #[error("Timestamp too old")]
-        Expired,
-
-        #[error("Time error: {0}")]
-        Time(#[from] std::time::SystemTimeError),
-    }
-
-    let (x5c, safetynet_response) =
-        |token: &str| -> Result<(Vec<x509::X509>, SafteyNetAttestResponse), SafetyNetError> {
-            trace!(?token);
-            use std::str::FromStr;
-            let jwsu = JwsCompact::from_str(token)?;
-
-            let certs = jwsu
-                .get_x5c_chain()?
-                .ok_or(SafetyNetError::MissingCertChain)?;
-
-            let leaf_cert = certs.first().cloned().ok_or(SafetyNetError::BadCert)?;
-
-            // Verify with the internal certificate.
-            let verifier = JwsX509Verifier::from_x509(leaf_cert.clone())
-                .map_err(|_| SafetyNetError::BadCert)?;
-
-            let verified_claims: SafteyNetAttestResponse = verifier
-                .verify(&jwsu)
-                .and_then(|jws| jws.from_json().map_err(|_| compact_jwt::JwtError::Serde))?;
-
-            // 3. Verify that the nonce attribute in the payload of response is identical to the Base64 encoding of the SHA-256 hash of the concatenation of authenticatorData and clientDataHash.
-            if verified_claims.nonce != data_to_verify.as_slice() {
-                return Err(SafetyNetError::NonceMismatch);
-            }
-
-            // 4. Verify that the SafetyNet response actually came from the SafetyNet service by following the steps in the SafetyNet online documentation.
-            let common_name = {
-                let name = leaf_cert
-                    .subject_name()
-                    .entries_by_nid(openssl::nid::Nid::COMMONNAME)
-                    .next()
-                    .ok_or(SafetyNetError::InvalidHostname)?;
-                name.data().as_utf8()?.to_string()
-            };
-
-            // §8.5.5 Verify that attestationCert is issued to the hostname "attest.android.com"
-            if common_name.as_str() != "attest.android.com" {
-                return Err(SafetyNetError::InvalidHostname);
-            }
-
-            // §8.5.6 Verify that the ctsProfileMatch attribute in the payload of response is true.
-            if !verified_claims.cts_profile_match {
-                return Err(SafetyNetError::CtsProfileMatchFailed);
-            }
-
-            // Verify sanity of timestamp in the payload
-            if !danger_ignore_timestamp {
-                let expires: std::time::Duration = std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)?
-                    + std::time::Duration::from_secs(60);
-                if verified_claims.timestamp_ms as u128 > expires.as_millis() {
-                    return Err(SafetyNetError::Expired);
-                }
-            }
-
-            Ok((certs, verified_claims))
-        }(response_str)
-        .map_err(|e| {
-            error!("jwt saftey-net error: {:?}", e);
-            WebauthnError::AttestationStatementResponseInvalid
-        })?;
-
-    let SafteyNetAttestResponse {
-        timestamp_ms: _,
-        nonce: _,
-        apk_package_name,
-        apk_certificate_digest_sha256,
-        cts_profile_match,
-        basic_integrity,
-        evaluation_type,
-    } = safetynet_response;
-
-    let metadata = AttestationMetadata::AndroidSafetyNet {
-        apk_package_name,
-        apk_certificate_digest_sha256,
-        cts_profile_match,
-        basic_integrity,
-        evaluation_type,
-    };
-
-    // 5. If successful, return implementation-specific values representing attestation type Anonymous CA and attestation trust path x5c.
-    Ok((ParsedAttestationData::Basic(x5c), metadata))
 }
 
 /// Verify the attestation chain
